requirements.txt
# Core framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation and serialization
pydantic[email]==2.5.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Visualization
plotly==5.17.0

# Configuration and environment
python-dotenv==1.0.0

# HTTP requests for integrations
requests==2.31.0

# Development dependencies
faker==20.1.0
pytest==7.4.3